requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for API calls
httpx==0.25.2
aiofiles==23.2.1

# Audio processing
python-speech-recognition==3.10.0
pydub==0.25.1